backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.3.0
firebase-admin==6.5.0
celery==5.4.0
redis==5.0.4
django-filter==24.2
drf-spectacular==0.27.2
